removed commented code, extracted the base class for write attribute handlers
