+++ threads/src/main/java/org/jboss/as/threads/ThreadsExtension.java
@@ -96,12 +96,6 @@ public class ThreadsExtension implements Extension {
                 THREAD_FACTORY_DESC);
         threadFactories.registerOperationHandler(ADD, ThreadFactoryAdd.INSTANCE, ThreadFactoryAdd.INSTANCE, false);
         threadFactories.registerOperationHandler(REMOVE, ThreadFactoryRemove.INSTANCE, ThreadFactoryRemove.INSTANCE, false);
-/*        threadFactories.registerReadWriteAttribute(THREAD_NAME_PATTERN, null, ThreadFactoryThreadNamePatternUpdate.INSTANCE,
-                AttributeAccess.Storage.CONFIGURATION);
-        threadFactories.registerReadWriteAttribute(GROUP_NAME, null, ThreadFactoryGroupNameUpdate.INSTANCE,
-                AttributeAccess.Storage.CONFIGURATION);
-        threadFactories.registerReadWriteAttribute(PRIORITY, null, ThreadFactoryPriorityUpdate.INSTANCE, AttributeAccess.Storage.CONFIGURATION);
-*/
         ThreadFactoryWriteAttributeHandler.INSTANCE.registerAttributes(threadFactories);
 
         final ManagementResourceRegistration boundedQueueThreadPools = subsystem.registerSubModel(
